Update responsive styles for smaller devices and improve layout
Adjust the media query in style.css to target devices with a maximum width of 1107px instead of 768px. This ensures that the website's layout is optimized for a wider range of smaller devices.

Also, update the margin-top values for certain elements to improve spacing and alignment on smaller screens. Specifically, increase the margin-top value for the h5 element inside the .flex1 class to 30px, and set the margin-top value for the .list1 class to 20px. Additionally, adjust the margin-top value for the h2 element inside the .end class to 100px, providing more space at the top and bottom of the element.

These changes enhance the responsiveness and overall visual experience of the website on smaller devices.